process_10xgenomics.py: substantial update to move most functions into libraries.
